chore: move oracle navigation to hook, move or delete associated functions
- removed passing down a function from a hook
- moved select oracle function into a hook that uses all necessary data from hooks
