+++ packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/helpers/index.ts
@@ -1,1 +1,2 @@
-export * from './handle-oracle-navigation';
+export * from './is-hcaptcha-oracle';
+export * from './should-navigate-to-registration';
