--- packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/helpers/navigate-to-hcaptcha-page.ts
@@ -1,18 +1,0 @@
-import { type NavigateFunction } from 'react-router-dom';
-import { routerPaths } from '@/router/router-paths';
-
-interface NavigateToHCaptchaPageProps {
-  siteKey: string | null | undefined;
-  navigate: NavigateFunction;
-}
-
-export const navigateToHCaptchaPage = ({
-  siteKey,
-  navigate,
-}: NavigateToHCaptchaPageProps): void => {
-  const route = siteKey
-    ? routerPaths.worker.HcaptchaLabeling
-    : routerPaths.worker.enableLabeler;
-
-  navigate(route);
-};
