--- packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/helpers/navigate-to-oracle-jobs-page.ts
@@ -1,12 +1,0 @@
-import { type NavigateFunction } from 'react-router-dom';
-import { routerPaths } from '@/router/router-paths';
-import { type Oracle } from '../hooks';
-
-export const navigateToOracleJobsPage = (
-  oracle: Oracle,
-  navigate: NavigateFunction
-): void => {
-  navigate(`${routerPaths.worker.jobs}/${oracle.address}`, {
-    state: { oracle },
-  });
-};
