--- packages/apps/human-app/frontend/src/modules/worker/jobs-discovery/helpers/navigate-to-registration-page.ts
@@ -1,11 +1,0 @@
-import { type NavigateFunction } from 'react-router-dom';
-import { routerPaths } from '@/router/router-paths';
-
-export const navigateToRegistrationPage = (
-  oracleAddress: string,
-  navigate: NavigateFunction
-): void => {
-  navigate(
-    `${routerPaths.worker.registrationInExchangeOracle}/${oracleAddress}`
-  );
-};
